LibWeb: Run changing-navigable history jobs on the history executor
Move the changing navigable job, its population resumption, and its
continuation application from LocalTraversableNavigable to
HistoryExecutor, together with the retained continuation state, which
is private to the executor again. These algorithms only used the
traversable for its page and its operation records.

The WebContent handlers for these jobs address the page's executor, as
does the base PageClient's history population completion.
