+++ Libraries/LibWeb/Page/Page.cpp
@@ -1462,7 +1462,7 @@ void Page::set_viewport_is_fullscreen(ViewportIsFullscreen is_fullscreen)
 
 void PageClient::history_navigation_params_creation_finished(HTML::CrossProcessId operation_id, HTML::HistoryNavigationPopulation population)
 {
-    as<HTML::LocalTraversableNavigable>(*page().local_root_navigable()).resume_history_navigation_population(operation_id, move(population));
+    page().history_executor().resume_history_navigation_population(operation_id, move(population));
 }
 
 void PageClient::request_navigation_start(HTML::LocalNavigable& navigable, NavigationTarget target, URL::URL const&, Utf16String navigation_id, Optional<HTML::NavigationStartRequest> start_request)
